refactor[#52]: 상세루틴 API 분리
