+++ src/api/routine/group/routines.ts
@@ -9,14 +9,6 @@ import {
   UpdateGroupRoutineResponse,
   DeleteGroupRoutineResponse,
   JoinGroupRoutineResponse,
-  CreateGroupRoutineDetailRequest,
-  CreateGroupRoutineDetailResponse,
-  UpdateGroupRoutineDetailRequest,
-  UpdateGroupRoutineDetailResponse,
-  DeleteGroupRoutineDetailResponse,
-  GroupRoutineDetailResponse,
-  UpdateGroupRoutineStatusRequest,
-  UpdateGroupRoutineStatusResponse,
 } from '../../../types/api';
 
 // 단체루틴 생성 API
@@ -26,67 +18,6 @@ export const createGroupRoutine = async (
   const response = await apiClient.post<
     ApiResponse<CreateGroupRoutineResponse>
   >('/api/v1/routines/groups', data);
-
-  return response.data;
-};
-
-// 단체루틴 상세 생성 API
-export const createGroupRoutineDetail = async (
-  groupRoutineListId: string,
-  data: CreateGroupRoutineDetailRequest,
-): Promise<ApiResponse<CreateGroupRoutineDetailResponse>> => {
-  const response = await apiClient.post<
-    ApiResponse<CreateGroupRoutineDetailResponse>
-  >(`/api/v1/routines/groups/${groupRoutineListId}/sub-routines`, data);
-
-  return response.data;
-};
-
-// 단체루틴 상세 수정 API
-export const updateGroupRoutineDetail = async (
-  groupRoutineListId: string,
-  data: UpdateGroupRoutineDetailRequest,
-): Promise<ApiResponse<UpdateGroupRoutineDetailResponse>> => {
-  const response = await apiClient.put<
-    ApiResponse<UpdateGroupRoutineDetailResponse>
-  >(`/api/v1/routines/groups/${groupRoutineListId}/sub-routines`, data);
-
-  return response.data;
-};
-
-// 단체루틴 상세 삭제 API
-export const deleteGroupRoutineDetail = async (
-  groupRoutineListId: string,
-  routineId: string,
-): Promise<ApiResponse<DeleteGroupRoutineDetailResponse>> => {
-  const response = await apiClient.delete<
-    ApiResponse<DeleteGroupRoutineDetailResponse>
-  >(`/api/v1/routines/groups/${groupRoutineListId}/sub-routines/${routineId}`);
-
-  return response.data;
-};
-
-// 단체루틴 상세 조회 API
-export const getGroupRoutineDetail = async (
-  groupRoutineListId: string,
-  groupRoutineId: string,
-): Promise<ApiResponse<GroupRoutineDetailResponse>> => {
-  const response = await apiClient.get<ApiResponse<GroupRoutineDetailResponse>>(
-    `/api/v1/routines/groups/${groupRoutineListId}/routines/${groupRoutineId}`,
-  );
-
-  return response.data;
-};
-
-// 단체루틴 상세루틴 성공/실패 API
-export const updateGroupRoutineStatus = async (
-  groupRoutineListId: string,
-  routineId: string,
-  data: UpdateGroupRoutineStatusRequest,
-): Promise<ApiResponse<UpdateGroupRoutineStatusResponse>> => {
-  const response = await apiClient.patch<
-    ApiResponse<UpdateGroupRoutineStatusResponse>
-  >(`/api/v1/routines/groups/${groupRoutineListId}/status/${routineId}`, data);
 
   return response.data;
 };
